meson: fix build with dakara_check subproject
the dakara_check library is not in the library path when calling
karaberus openapi or the tests.
We could continue with the workaround in a881c5e and keep linking to
dakara_check, and add it to the library path... but at this point we're
hurting ourselves more than we should.
karaberus_tools can be the interface to all our dependent C libraries
and that will not bite us in the ass later.