meson: force compilation of karaberus_tools dependencies
The libkaraberus_tools target doesn't need to the objects of its
dependencies to be built.
Generally when building an executable this wouldn't be an issue but
since we are working with a custom target it can fail in our case.
So now libkaraberus_tools depends on the objects of dakara_check and its
dependencies when built as a subproject.