server: implement dakara feeding/sync
Feed tags/medias/artists/karas to dakara on upload and on startup.
Define tag colors (hue) so we can set colors in karaberus and have it
change in dakara at the same time.

Could be obviously improved by checking locally for changes of karaokes,
medias and artists instead of just PUTing everything on sync. Probably
not worth it implement that for tags. We can't yet update work types
decently because dakara doesn't return their IDs.
(https://github.com/DakaraProject/dakara-server/pull/165)

This assumes that video files will all be called `{id}.mkv`, so the
downloader will have to follow this naming scheme to work with dakara.
The files are not all matroska files, but it should not be a problem in
practice.

#40