server: add endpoint to refresh mugen imports
metadata imports were meant to be a one time operation, but there could
be reasons to want to update (including fixes to the import) so it makes
sense to have a way to reimport everything.
we probably need to add a field to say that an imported karaoke should
never be updated by this operation.