server: actually streaming uploads
Sometimes you see that your compression middleware sets a Content-Length
and you end up calling the cops because the StreamingResponse wasn't
streaming.

As it turns out fiber.Compress is not at fault because it can compress a
stream if that's what it gets, in which case it doesn't set the
Content-Length header. (That's what happens after this commit)

The BodyWriter in our cases returns a *fiber.Ctx which does have the
Write function that we used previously. Which appends each buffer that
we give to fiber's internal buffer (actually fasthttp's).
But for streaming we actually want to call SendStream, which as the name
implies will send a stream.

We can't test the download endpoints as we used to but at least the
server doesn't eat your RAM now.