meson|ci: build with go buildmode=c-archive
go builds a .a file that we can link against in a C program.
It works well for us because then meson is aware that we are building an
executable linking against libraries that we build from subprojects
(which was always a complicated thing to tell to meson with the previous
build)

Sadly it seems to miscompile (segfaults on launch) when built with musl
so I ended up building it from the github runner images and put the
results in an image.