server: only return tokens on creation
We show full tokens in /api/token which allows for easy privilege
escalation due to another bug on the definition of the route.
When in fact it is not necessary for general use, most APIs show the
token once and if you lost a token you can create a new one (and delete
the previous token).