relicense copyimglibs
it’s not really meant as a part of the license of karaberus and could be
useful in other projects to create images.
I don’t see much point licensing the script considering how it’s used
but it should be possible to copy it in another project and use it in
the same way.